+++ pkgs/native_assets_cli/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.2.0
 
-- *Breaking change* Rename `Asset.name` to `Asset.id`
+- **Breaking change** Rename `Asset.name` to `Asset.id`
   ([#100](https://github.com/dart-lang/native/issues/100)).
 - Added topics.
 - Fixed metadata example.